Add Dockerfiles for using of conda krb5
Using the krb5 1.19.1 of recent conda greatly simplifies jupyter (and
perspectively dask-gateway) image creation. But currently this krb5 is
too current for our Debian testing-based gssproxy image which only has
krb5 1.18.3. This causes the gssproxy to reject requests of the client
module (proxymech) as invalid.

Fedora 34 has krb5 1.19.1 and running its gssproxy on top of those makes
the interaction work again. This is meant as a PoC and therefore placed
in separate Dockerfiles for now. For maintainability the gssproxy image
should likely be based on the conda packages.