Add acquisition of oauth credentials for git
Both Github and Gitlab allow authentication using an OAuth2 access token
when accessing git repos via HTTPS. As a demonstrator, add access token
acquisition for Gitlab to the Jupyterhub login process by way of some
additional redirects and handlers.

Jupyterhub auth state is enabled as a means to save and transport those
credentials between requests and requires a custom encryption key to
be configured.

A .git-credentials file for the "store" git credential helper is saved
in a user-specific k8s secret and mounted into the singleuser notebook
container. A lifecycle command hook configures the store credential
helper to be used.