Split out custom python code
Move custom credential retrieval code to pure python files. This allows
for better accessibility and easier development such as syntax checking
and linting. To implement, we require the jupyter_config.d drop file
inclusion logic present in recent zero-to-jupyterhub helm charts.
Therefore to upgrade to daskhub 2021.8.1. To stay in sync with the new
charts, we switch to a current version of the jupyterhub container
image.